project/requirements.txt
Django==2.1.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-mptt==0.9.1
django-parler==1.9.2
django-widget-tweaks==1.4.3
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
